refactor: decouple auth_manager and pg_catalog
